convert-cli: move encode back to cmd
